ina3221

+++ drivers/sensor/ti/ina3221/ina3221.h
@@ -69,7 +69,7 @@ struct ina3221_config {
 	enum ina3221_conv_time conv_time_bus;
 	enum ina3221_conv_time conv_time_shunt;
 	bool enable_channel[3];
-	uint16_t shunt_r[3];
+	uint32_t shunt_r[3];
 };
 
 struct ina3221_data {
